Refactored the TableNotFoundException in same package name
