+++ encoders/src/main/scala/org/apache/spark/sql/TableNotFoundException.scala
@@ -1,0 +1,24 @@
+/*
+ * Copyright (c) 2018 SnappyData, Inc. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you
+ * may not use this file except in compliance with the License. You
+ * may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+ * implied. See the License for the specific language governing
+ * permissions and limitations under the License. See accompanying
+ */
+package org.apache.spark.sql
+
+class TableNotFoundException(schema: String, table: String, cause: Option[Throwable] = None)
+  extends AnalysisException(s"Table or view '$table' not found in schema '$schema'",
+    cause = cause)
+
+
+class PolicyNotFoundException(schema: String, name: String, cause: Option[Throwable] = None)
+  extends AnalysisException(s"Policy '$name' not found in schema '$schema'", cause = cause)
